--- src/constants/qurey-key.ts
@@ -1,3 +1,0 @@
-export const QUERY_KEY = Object.freeze({
-  //리펙터링하다 키는 삭제했지만 해당 상수obj는 사용할 예정이라 남겨두었습니다.
-});
